First cut at object browser pagination support.
While this does work, the fetching logic needs to be moved out of
DS3BrowserItem and into DS3BrowserModel so it can make use of
beginInsertRows and endInsertRows when inserting new objects into the
model.  That will come after this.

If there are more objects to load, a "Click to load more" page break row is
rendered, just like in the BlackPearl GUI.  Clicking it does exactly that.
The way we get the page break row to span all columns is currently a hack since the model now
knows about the view.  A better way would be to have the DS3Browser
view call QTreeView.isFirstColumnSpanned but I'm not sure where that
should be done.  It's easy to have the model instead call it since
there's already a good spot to do that.  A better way might be to
override QTreeView.drawRow and call it in there if it's rendering a page
break row.  That way, the model wouldn't have to know anything about the
view.