Update the lower-level portions of the app so it can report bulk put job progress updates to the GUI.  The GUI half currently only logs a debug messages when it receives a job progress update but that will change. This should be almost everything needed to do things like update a progress bar as one or more objects are being PUT.  Still needs some sort of job state though.
The Client emits a Job object via a Qt signal everytime the C SDK "read file"
callback is called.  This Job object is meant to be a higher-level
version of the lower-level BulkWorkItem and suitable for a GUI to use to
update progress bars, etc.  This signal is connected to a Qt slot in the
DS3Browser that will eventually do the GUI side of the job progress
update.

I considered sending a BulkWorkItem pointer through the signal/slot
instead of introducing a new Job class, however, the WorkItem classes
were really meant for Client only.  Sending a new Job object, by value,
through the signal/slot seemed easier and more appropriate although Job
doesn't have access to everything the original WorkItem does.  This
could be troublesome in the future depending on how much data we want to
report to the user.