Checkpoint for adding support for bulk get's via drag/drop operations. This will make the "bulk get" request(s) and attempt to parse the response but not actually get any objects.  Dragging objects from multiple buckets at the same time will result in multiple bulk get requests.
The Client's bulk get code path is modeled after the bulk put path.  It
currently parses the bulk response's "list of lists" the same way that bulk
puts are done, which is wrong; it incorrectly parses all lists in parallel.
This will be fixed separate checkin.  A future refacter will help
eliminate the code duplication between the bulk get/put code paths.

Lots of refactoring of the BulkWorkItem classes so BulkGetWorkItem can
share code with BulkPutWorkItem.