Authenticate the session before closing the new session dialog so users have the opportunity to, for example, fix an incorrect secret key. Before, they had to restart the app to try again.
To authenticate the session, a "GET bucket" request is made for a bucket
that should never exist (but it should work if for some reason the bucket
does exist).  A 404 error (not found) means it's authenticated.  This
type of request was used since it's quick and shouldn't result in a large
payload.  We could switch it to a HEAD request in the future.  The
downside is that our server logs will have logs for requests that have
technically failed, however, the name of the bucket we're trying to "GET" is
self documenting and clearly indicates why the request is being made.

Another way to do this would be to move the initial "GET server" request
to the SessionDialog class but that's not very clean.